Rename migration files from v195 to v196
Renamed migration classes and utilities to v196 for MySQL and PostgreSQL.
